fcfs_single basic build theoretical
The original files don't compile for me without a bunch of errors, but I
have a theoretical version of the fcfs single core, although I haven't
yet been able to test it. Due to the weird errors I made a separate
process struct for this, but I'll merge them together when it's all
working.